Deductions and capital gains
