+++ openfisca_us/variables/irs/outputs.py
@@ -288,20 +288,6 @@ class c03260(Variable):
         )
 
 
-class c04470(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Itemized deductions after phase-out"
-    unit = USD
-    documentation = (
-        "Itemized deductions after phase-out (zero for non-itemizers)"
-    )
-
-    def formula(tax_unit, period, parameters):
-        return max_(0, tax_unit("c21060", period) - tax_unit("c21040", period))
-
-
 class c05700(Variable):
     value_type = float
     entity = TaxUnit
@@ -310,45 +296,6 @@ class c05700(Variable):
         "search taxcalc/calcfunctions.py for how calculated and used"
     )
     unit = USD
-
-
-class c05800(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Income tax before credits"
-    unit = USD
-    documentation = "Total (regular + AMT) income tax liability before credits (equals taxbc plus c09600)"
-
-    def formula(tax_unit, period, parameters):
-        return add(
-            tax_unit,
-            period,
-            ["regular_tax_before_credits", "alternative_minimum_tax"],
-        )
-
-
-income_tax_before_credits = variable_alias("income_tax_before_credits", c05800)
-
-
-class c07100(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Income tax non-refundable credits"
-    documentation = (
-        "Total non-refundable credits used to reduce positive tax liability"
-    )
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        credits = parameters(period).irs.credits.non_refundable
-        return add(tax_unit, period, credits)
-
-
-income_tax_non_refundable_credits = variable_alias(
-    "income_tax_non_refundable_credits", c07100
-)
 
 
 class c07240(Variable):
@@ -411,189 +358,6 @@ class c08000(Variable):
     unit = USD
 
 
-class c09200(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    unit = USD
-    label = "Income tax before refundable credits"
-    documentation = "Income tax liability (including othertaxes) after non-refundable credits are used, but before refundable credits are applied"
-
-    def formula(tax_unit, period, parameters):
-        return max_(
-            0,
-            tax_unit("income_tax_before_credits", period)
-            - tax_unit("income_tax_non_refundable_credits", period),
-        )
-
-
-income_tax_before_refundable_credits = variable_alias(
-    "income_tax_before_refundable_credits", c09200
-)
-
-
-class c17000(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Medical expense deduction"
-    unit = USD
-    documentation = "Sch A: Medical expenses deducted (component of pre-limitation c21060 total)"
-
-    def formula(tax_unit, period, parameters):
-        medical = parameters(period).irs.deductions.itemized.medical
-        has_aged = (tax_unit("age_head", period) >= 65) | (
-            tax_unit("tax_unit_is_joint", period)
-            & (tax_unit("age_spouse", period) >= 65)
-        )
-        medical_floor_ratio = (
-            medical.floor.base + has_aged * medical.floor.aged_addition
-        )
-        medical_floor = medical_floor_ratio * max_(
-            tax_unit("c00100", period), 0
-        )
-        return max_(
-            0,
-            tax_unit("filer_e17500", period) - medical_floor,
-        )
-
-
-class c18300(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "SALT deduction"
-    unit = USD
-    documentation = "Sch A: State and local taxes plus real estate taxes deducted (component of pre-limitation c21060 total)"
-
-    def formula(tax_unit, period, parameters):
-        c18400 = max_(tax_unit("filer_e18400", period), 0)
-        c18500 = tax_unit("filer_e18500", period)
-        salt = parameters(period).irs.deductions.itemized.salt_and_real_estate
-        cap = salt.cap[tax_unit("mars", period)]
-        return min_(c18400 + c18500, cap)
-
-
-class c19200(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Interest deduction"
-    unit = USD
-    documentation = (
-        "Sch A: Interest deducted (component of pre-limitation c21060 total)"
-    )
-
-    def formula(tax_unit, period, parameters):
-        return tax_unit("filer_e19200", period)
-
-
-class c19700(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Charitable deduction"
-    unit = USD
-    documentation = "Sch A: Charity contributions deducted (component of pre-limitation c21060 total)"
-
-    def formula(tax_unit, period, parameters):
-        charity = parameters(period).irs.deductions.itemized.charity
-        posagi = tax_unit("posagi", period)
-        lim30 = min_(
-            charity.ceiling.non_cash * posagi,
-            tax_unit("filer_e20100", period),
-        )
-        c19700 = min_(
-            charity.ceiling.all * posagi,
-            lim30 + tax_unit("filer_e19800", period),
-        )
-        return max_(c19700, 0)
-
-
-class c20500(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Casualty deduction"
-    unit = USD
-    documentation = "Sch A: Net casualty or theft loss deducted (component of pre-limitation c21060 total)"
-
-    def formula(tax_unit, period, parameters):
-        casualty = parameters(period).irs.deductions.itemized.casualty
-        floor = casualty.floor * tax_unit("posagi", period)
-        deduction = max_(0, tax_unit("filer_g20500", period) - floor)
-        return deduction * (1 - casualty.haircut)
-
-
-class c20800(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Miscellaneous deductions"
-    unit = USD
-    documentation = "Sch A: Net limited miscellaneous deductions deducted (component of pre-limitation c21060 total)"
-
-    def formula(tax_unit, period, parameters):
-        misc = parameters(period).irs.deductions.itemized.misc
-        floor = misc.floor * tax_unit("posagi", period)
-        deduction = max_(0, tax_unit("filer_e20400", period) - floor)
-        return deduction * (1 - misc.haircut)
-
-
-class c21040(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Phased-out itemized deductions"
-    unit = USD
-    documentation = "Itemized deductions that are phased out"
-
-    def formula(tax_unit, period, parameters):
-        nonlimited = add(tax_unit, period, ["c17000", "c20500"])
-        phaseout = parameters(period).irs.deductions.itemized.phaseout
-        mars = tax_unit("mars", period)
-        c21060 = tax_unit("c21060", period)
-        phaseout_amount_cap = phaseout.cap * max_(0, c21060 - nonlimited)
-        uncapped_phaseout = max_(
-            0,
-            (
-                (tax_unit("posagi", period) - phaseout.start[mars])
-                * phaseout.rate
-            ),
-        )
-        return min_(
-            uncapped_phaseout,
-            phaseout_amount_cap,
-        )
-
-
-class c21060(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Gross itemized deductions"
-    unit = USD
-    documentation = (
-        "Itemized deductions before phase-out (zero for non-itemizers)"
-    )
-
-    def formula(tax_unit, period, parameters):
-        ELEMENTS = ["c17000", "c18300", "c19200", "c19700", "c20500", "c20800"]
-        return add(tax_unit, period, ELEMENTS)
-
-
-class c23650(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Net capital gains"
-    unit = USD
-    documentation = "Net capital gains (long and short term) before exclusion"
-
-    def formula(tax_unit, period, parameters):
-        return add(tax_unit, period, ["filer_p23250", "filer_p22250"])
-
-
 class tax_unit_is_joint(Variable):
     value_type = bool
     entity = TaxUnit
@@ -604,64 +368,6 @@ class tax_unit_is_joint(Variable):
     def formula(tax_unit, period, parameters):
         mars = tax_unit("mars", period)
         return mars == mars.possible_values.JOINT
-
-
-class c62100(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "AMT taxable income"
-    unit = USD
-    documentation = "Alternative Minimum Tax (AMT) taxable income"
-
-    def formula(tax_unit, period, parameters):
-        # Form 6251, Part I
-        c00100 = tax_unit("c00100", period)
-        e00700 = tax_unit("filer_e00700", period)
-        c62100_if_no_standard = (
-            c00100
-            - e00700
-            - tax_unit("c04470", period)
-            + max_(
-                0,
-                min_(
-                    tax_unit("c17000", period),
-                    0.025 * c00100,
-                ),
-            )
-            + tax_unit("c18300", period)
-            + tax_unit("c20800", period)
-            - tax_unit("c21040", period)
-        )
-        c62100 = where(
-            tax_unit("standard", period) == 0,
-            c62100_if_no_standard,
-            c00100 - e00700,
-        ) + tax_unit(
-            "filer_cmbtp", period
-        )  # add income not in AGI but considered income for AMT
-        amt = parameters(period).irs.income.amt
-        mars = tax_unit("mars", period)
-        separate_addition = max_(
-            0,
-            min_(
-                amt.exemption.amount[mars],
-                amt.exemption.phaseout.rate
-                * (c62100 - amt.exemption.separate_limit),
-            ),
-        ) * (mars == mars.possible_values.SEPARATE)
-        return c62100 + separate_addition
-
-
-class c87668(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    unit = USD
-    documentation = "American Opportunity Credit non-refundable amount from Form 8863 (included in c07230)"
-
-    def formula(tax_unit, period, parameters):
-        return tax_unit("non_refundable_american_opportunity_credit", period)
 
 
 class care_deduction(Variable):
@@ -702,126 +408,6 @@ class charity_credit(Variable):
     unit = USD
 
 
-class dwks6(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "DWKS6"
-    unit = USD
-    definition_period = YEAR
-
-    def formula(tax_unit, period, parameters):
-        dwks2 = tax_unit("filer_e00650", period)
-        dwks3 = tax_unit("filer_e58990", period)
-        # dwks4 always assumed to be zero
-        dwks5 = max_(0, dwks3)
-        return max_(0, dwks2 - dwks5)
-
-
-class dwks9(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "DWKS9"
-    unit = USD
-    definition_period = YEAR
-
-    def formula(tax_unit, period, parameters):
-        p23250 = tax_unit("filer_p23250", period)
-        c23650 = tax_unit("c23650", period)
-        dwks7 = min_(p23250, c23650)  # SchD lines 15 and 16, respectively
-        # dwks8 = min(dwks3, dwks4)
-        # dwks9 = max(0., dwks7 - dwks8)
-        # BELOW TWO STATEMENTS ARE UNCLEAR IN LIGHT OF dwks9=... COMMENT
-        e01100 = tax_unit("filer_e01100", period)
-        c24510 = where(e01100 > 0, e01100, max_(0, dwks7) + e01100)
-        return max_(0, c24510 - min_(0, tax_unit("filer_e58990", period)))
-
-
-class dwks10(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "DWKS10"
-    documentation = (
-        "search taxcalc/calcfunctions.py for how calculated and used"
-    )
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        dwks10_if_gains = add(tax_unit, period, ["dwks6", "dwks9"])
-        dwks10_if_no_gains = max_(
-            0,
-            min_(
-                tax_unit("filer_p23250", period),
-                tax_unit("c23650", period),
-            ),
-        ) + tax_unit("filer_e01100", period)
-        return where(
-            tax_unit("hasqdivltcg", period),
-            dwks10_if_gains,
-            dwks10_if_no_gains,
-        )
-
-
-class dwks13(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "DWKS13"
-    unit = USD
-    documentation = (
-        "search taxcalc/calcfunctions.py for how calculated and used"
-    )
-
-    def formula(tax_unit, period, parameters):
-        dwks1 = tax_unit("c04800", period)
-        e24515 = tax_unit("filer_e24515", period)
-        dwks11 = e24515 + tax_unit(
-            "filer_e24518", period
-        )  # Sch D lines 18 and 19, respectively
-        dwks9 = tax_unit("dwks9", period)
-        dwks12 = min_(dwks9, dwks11)
-        dwks10 = tax_unit("dwks10", period)
-        return (dwks10 - dwks12) * tax_unit("hasqdivltcg", period)
-
-
-class dwks14(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "DWKS14"
-    unit = USD
-    documentation = (
-        "search taxcalc/calcfunctions.py for how calculated and used"
-    )
-
-    def formula(tax_unit, period, parameters):
-        dwks1 = tax_unit("c04800", period)
-        dwks13 = tax_unit("dwks13", period)
-        return max_(0, dwks1 - dwks13) * tax_unit("hasqdivltcg", period)
-
-
-class dwks19(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "DWKS14"
-    unit = USD
-    documentation = (
-        "search taxcalc/calcfunctions.py for how calculated and used"
-    )
-
-    def formula(tax_unit, period, parameters):
-        dwks14 = tax_unit("dwks14", period)
-        capital_gains = parameters(period).irs.capital_gains.brackets
-        mars = tax_unit("mars", period)
-        dwks1 = tax_unit("c04800", period)
-        dwks16 = min_(capital_gains.thresholds["1"][mars], dwks1)
-        dwks17 = min_(dwks14, dwks16)
-        dwks10 = tax_unit("dwks10", period)
-        dwks18 = max_(0, dwks1 - dwks10)
-        return max_(dwks17, dwks18) * tax_unit("hasqdivltcg", period)
-
-
 class fstax(Variable):
     value_type = float
     entity = TaxUnit
@@ -830,41 +416,6 @@ class fstax(Variable):
         "search taxcalc/calcfunctions.py for how calculated and used"
     )
     unit = USD
-
-
-class invinc_agi_ec(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "Exclusion of investment income from AGI"
-    unit = USD
-    documentation = (
-        "Always equal to zero (will be removed in a future version)"
-    )
-    definition_period = YEAR
-
-
-class invinc_ec_base(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "AGI investment income exclusion"
-    unit = USD
-    documentation = "Exclusion of investment income from AGI"
-
-    def formula(tax_unit, period, parameters):
-        # Limitation on net short-term and
-        # long-term capital losses
-        limited_capital_gain = max_(
-            -3000.0 / tax_unit("sep", period),
-            add(tax_unit, period, ["filer_p22250", "filer_p23250"]),
-        )
-        OTHER_INV_INCOME_VARS = ["e00300", "e00600", "e01100", "e01200"]
-        other_inv_income = add(
-            tax_unit,
-            period,
-            ["filer_" + variable for variable in OTHER_INV_INCOME_VARS],
-        )
-        return limited_capital_gain + other_inv_income
 
 
 class filer_setax(Variable):
